rename types progress chart
